requirements.txt
google-genai>=0.2.0
requests>=2.25.0
python-dotenv>=0.19.0
gradio>=4.0.0
Pillow>=8.0.0
